+++ src/pages/SignUp/index.tsx
@@ -22,17 +22,23 @@ const SignIn: React.FC = () => {
   async function handleSubmit(
     sign: Omit<ISignUp, "created_at" | "updated_at">
   ): Promise<void> {
-    try {
-      setLoading(true);
-      await api.post("/users", {
-        ...sign,
-      });
-      toast.success("Cadastro realizado com sucesso!");
-      history.push("/");
-    } catch (err) {
-      toast.error(err.response.data.message);
-      setLoading(false);
-      localStorage.clear();
+    if (!sign.email || !sign.password) {
+      toast.error("Preencha todos os campos para se cadastrar!");
+    } else if (sign.password.length < 8) {
+      toast.error("Sua senha deve conter 8 ou mais caracteres!");
+    } else {
+      try {
+        setLoading(true);
+        await api.post("/users", {
+          ...sign,
+        });
+        toast.success("Cadastro realizado com sucesso!");
+        history.push("/");
+      } catch (err) {
+        toast.error(err.response.data.message);
+        setLoading(false);
+        localStorage.clear();
+      }
     }
   }
 
@@ -40,7 +46,7 @@ const SignIn: React.FC = () => {
     <Container>
       <img src={logo} alt="logo" />
       <Form ref={formRef} onSubmit={handleSubmit}>
-        <Input name="email" placeholder="Seu e-mail" />
+        <Input type="email" name="email" placeholder="Seu e-mail" />
         <Input type="password" name="password" placeholder="Sua senha" />
         <button type="submit" data-testid="add-food-button">
           <p className="text">{loading ? "Carregando..." : "Cadastrar"}</p>
